% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytoMDS.R
\name{CytoMDS}
\alias{CytoMDS}
\title{CytoMDS package}
\description{
\code{CytoPipelineGUI} is the companion package of \code{CytoPipeline},
and is used for interactive visualization.
It implements two shiny applications :
\itemize{
\item a shiny app for interactive comparison of flow frames
that are the results of CytoProcessingSteps of the same
or different CytoPipeline experiments.
It is launched using the following statement: \code{CytoPipelineCheckApp()}
\item a shiny app for interactive visualization and manual adjustments of scale
transformation objects. It is launched using the following statement:
\code{ScaleTransformApp()}
}
}
