% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{getPairwiseEMDDist}
\alias{getPairwiseEMDDist}
\title{Calculate all pairwise Earth Mover's distances
between flowFrames of a flowSet}
\usage{
getPairwiseEMDDist(fs, channels = NULL, verbose = FALSE, ...)
}
\arguments{
\item{fs}{a flowCore::flowSet}

\item{channels}{which channels (integer index(ices) or character(s)):
\itemize{
\item if it is a character vector,
it can refer to either the channel names, or the marker names
\item if it is a numeric vector,
it refers to the indexes of channels in \code{fs}
\item if NULL all scatter and fluorescent channels of \code{fs}
will be selected
}}

\item{verbose}{if \code{TRUE}, output a message
after each single distance calculation}

\item{...}{additional parameters passed to \code{getEMDDist()}}
}
\value{
a distance matrix (full symmetric with 0. diagonal)
}
\description{
Calculate all pairwise Earth Mover's distances
between flowFrames of a flowSet
}
\examples{

library(CytoPipeline)

data(OMIP021Samples)

# estimate scale transformations 
# and transform the whole OMIP021Samples

transList <- estimateScaleTransforms(
    ff = OMIP021Samples[[1]],
    fluoMethod = "estimateLogicle",
    scatterMethod = "linearQuantile",
    scatterRefMarker = "BV785 - CD3")

OMIP021Trans <- CytoPipeline::applyScaleTransforms(
    OMIP021Samples, 
    transList)
    
# calculate pairwise distances using only FSC-A & SSC-A channels
pwDist <- getPairwiseEMDDist(
    fs = OMIP021Trans,
    channels = c("FSC-A", "SSC-A"))

}
