% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{pairwiseEMDDist}
\alias{pairwiseEMDDist}
\title{Calculate all pairwise Earth Mover's distances
between flowFrames of a flowSet}
\usage{
pairwiseEMDDist(
  nSamples,
  rowRange = c(1, nSamples),
  colRange = c(min(rowRange), nSamples),
  loadFlowFrameFUN,
  channels = NULL,
  verbose = FALSE,
  useBiocParallel = FALSE,
  BPPARAM = BiocParallel::bpparam(),
  memSize = Inf,
  ...
)
}
\arguments{
\item{channels}{which channels (integer index(ices) or character(s)):
\itemize{
\item if it is a character vector,
it can refer to either the channel names, or the marker names
\item if it is a numeric vector,
it refers to the indexes of channels in \code{fs}
\item if NULL all scatter and fluorescent channels of \code{fs}
will be selected
}}

\item{verbose}{if \code{TRUE}, output a message
after each single distance calculation}

\item{useBiocParallel}{if \code{TRUE}, use \code{BiocParallel} for computation of the
pairwise distances in parallel - one (i,j) at a time.
Note the \code{BiocParallel} function used internally is \code{bplapply()}}

\item{BPPARAM}{if \code{useBiocParallel} is TRUE, sets the \code{BPPARAM} back-end to
be used for the computation. If not provided, will use the top back-end on
the \code{BiocParallel::registered()} stack.}

\item{...}{additional parameters passed to \code{getEMDDist()}}

\item{fs}{a flowCore::flowSet}

\item{fs2}{a flowCore::flowSet
\itemize{
\item if fs2 is NULL, pairwise distances will be calculated for all pairs of \code{fs}
\item if fs2 is not NULL, distances will be calculated for all pairs where the
first element comes from \code{fs}, and the second element comes from \code{fs2}
}}

\item{nRowBlock}{(only with \code{useBiocParallel}) specifies the number of blocks
to divide the row of the matrix into. If set to \code{NULL}, will try to choose
it such that the number of tasks is near \code{BiocParallel::bpWorkers(BPPARAM)}}
}
\value{
a distance matrix of pairwise distances
(full symmetric with 0. diagonal)
}
\description{
Calculate all pairwise Earth Mover's distances
between flowFrames of a flowSet
}
\examples{

library(CytoPipeline)

data(OMIP021Samples)

# estimate scale transformations 
# and transform the whole OMIP021Samples

transList <- estimateScaleTransforms(
    ff = OMIP021Samples[[1]],
    fluoMethod = "estimateLogicle",
    scatterMethod = "linearQuantile",
    scatterRefMarker = "BV785 - CD3")

OMIP021Trans <- CytoPipeline::applyScaleTransforms(
    OMIP021Samples, 
    transList)
    
# calculate pairwise distances using only FSC-A & SSC-A channels
pwDist <- getPairwiseEMDDist(
    fs = OMIP021Trans,
    channels = c("FSC-A", "SSC-A"))

}
