% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytoMDS.R
\docType{package}
\name{CytoMDS}
\alias{CytoMDS}
\title{CytoMDS package}
\description{
\code{CytoMDS} implements a low dimensional visualization of a set of cytometry
samples, in order to visually assess the 'distances' between them.
This, in turn, can greatly help the user to identify quality issues
like batch effects or outlier samples, and/or check the presence of
potential sample clusters that might align with the experimental design.

The CytoMDS algorithm combines, on the one hand,
the concept of Earth Mover's Distance (EMD), a.k.a. Wasserstein metric
and, on the other hand, the Multi Dimensional Scaling (MDS) algorithm
for the low dimensional projection.

Also, the package provides some diagnostic tools for
both checking the quality of the MDS projection,
as well as tools to help with the interpretation of
the axes of the projection.
}
\seealso{
\link{ggplotSampleMDS}, \link{ggplotSampleMDSShepard}, \link{computeMetricMDS}
}
