% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistSum-class.R
\docType{class}
\name{DistSum-class}
\alias{DistSum-class}
\alias{DistSum,dist-method}
\alias{DistSum,list-method}
\alias{show,DistSum-method}
\alias{nDistr}
\alias{nFeatures}
\alias{featureNames}
\alias{featureNames<-}
\alias{distrNames}
\alias{distrNames<-}
\alias{getPWDist}
\title{DistSum class}
\usage{
\S4method{DistSum}{dist}(object)

\S4method{DistSum}{list}(object)

\S4method{show}{DistSum}(object)

nDistr(x)

nFeatures(x)

featureNames(x)

featureNames(x) <- value

distrNames(x)

distrNames(x) <- value

getPWDist(x, whichFeatures = NULL)
}
\arguments{
\item{object}{a \code{DistSum} object}

\item{x}{a \code{DistSum} object}

\item{value}{the new feature names to be assigned}

\item{whichFeatures}{either an array of feature names,
or an array of feature indices, or NULL
If NULL, the full disatnce (for all features) will be returned}
}
\value{
nothing
}
\description{
Class representing pairwise distances between multiple multidimensional
distributions, when the distance is calculated as a sum of marginal
distribution distances.
}
\section{Slots}{

\describe{
\item{\code{pwFullDist}}{An object of class \code{dist} storing the full pairwise
distance matrix. The \code{dist} object stores the triangular relevant part
of the symmetrix, 0 diagonal, matrix only.}

\item{\code{pwDistPerDim}}{A \code{list} of \code{dist} objects storing the contribution
of each dimension of the multidimensional distributions
to the full distance matrix}

\item{\code{featureNames}}{A \code{character} vector containing the feature names}

\item{\code{distrNames}}{A \code{character} vector containing the distribution names}
}}

\examples{

M <- matrix(data = rnorm(35), ncol = 7)
rownames(M) <- paste0("feat", 1:5)
D <- dist(M)




}
