% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplots.R
\name{ggplotSampleMDSShepard}
\alias{ggplotSampleMDSShepard}
\title{Plot of Metric MDS object - Shepard diagram}
\usage{
ggplotSampleMDSShepard(
  mdsObj,
  nDim,
  title = "Multi Dimensional Scaling - Shepard's diagram",
  pointSize = 0.5,
  displayPseudoRSq = TRUE
)
}
\arguments{
\item{mdsObj}{a MDS object calculated by the SMACOF algorithm using
the computeMetricMDS() function}

\item{nDim}{(optional) number of dimensions to use when calculating
Shepard's diagram and pseudoRSquare.
If missing, it will be set equal to the number of projection dimensions
as calculated in \code{mdsObj}}

\item{title}{title to give to the plot}

\item{pointSize}{plot size of points}

\item{displayPseudoRSq}{if TRUE, display pseudo RSquare in subtitle, on top
of nb of dimensions}
}
\value{
a ggplot object
}
\description{
\code{ggplotSampleMDSShepard} uses ggplot2
to provide plot of Metric MDS results.
Shepard diagram provides a scatter plot of :
\itemize{
\item on the x axis, the high dimensional pairwise distances
between each sample pairs
\item on the y axis, the corresponding pairwise distances in the obtained
low dimensional projection
}
}
\examples{

# prepare data, build MDS object
example("computeMetricMDS") 

# Shepard diagrams 

p2D <- ggplotSampleMDSShepard(
    mdsObj,
    nDim = 2,
    pointSize = 1,
    title = "Shepard with 2 dimensions")

p3D <- ggplotSampleMDSShepard(
    mdsObj,
    nDim = 3,
    title = "Shepard with 3 dimensions") 
    #' 
pDefD <- ggplotSampleMDSShepard(
    mdsObj,
    title = "Shepard with default nb of dimensions") 

}
\seealso{
\link{ggplotSampleMDS}, \link{computeMetricMDS}
}
