% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplots.R
\name{ggplotSampleMDSWrapBiplots}
\alias{ggplotSampleMDSWrapBiplots}
\title{SampleMDS biplot wrapping}
\usage{
ggplotSampleMDSWrapBiplots(
  mdsObj,
  extVariableList,
  ncol = NULL,
  nrow = NULL,
  byrow = NULL,
  ...
)
}
\arguments{
\item{mdsObj}{a MDS object calculated by the SMACOF algorithm using
the computeMetricMDS() function}

\item{extVariableList}{should be a named list of external variable matrices
Each element of the list should be a matrix with named columns
corresponding to the variables.
The number of rows should be the same as the number of samples.}

\item{ncol}{passed to \code{patchwork::wrap_plots()}}

\item{nrow}{passed to \code{patchwork::wrap_plots()}}

\item{byrow}{passed to \code{patchwork::wrap_plots()}}

\item{...}{additional parameters passed to \code{ggplotSampleMDS()}
(if used)}
}
\value{
a ggplot object
}
\description{
\code{ggplotSampleMDSWrapBiplots} calls \code{ggplotSampleMDS}
repeatly to generate biplots with different sets of external variables
and align them in a grid using the \code{patchwork} package, in a similar fashion
as \code{ggplot2::facet_wrap()} does.
}
\examples{

# prepare data, build MDS object
example("computeMetricMDS")

# plot mds projection on axes 1 and 2,
# use 'group' for colour, 'type' for shape, and no label 

p_12 <- ggplotSampleMDS(
    mdsObj = mdsObj,
    pData = flowCore::pData(fsAll),
    projectionAxes = c(1,2),
    pDataForColour = "grpId",
    pDataForShape = "type")

# try to associate axes with median or std deviation of each channel
# => use bi-plots

extVarList <- channelSummaryStats(
    fsAll,
    channels = c("FSC-A", "SSC-A"),
    statFUNs = c("median" = stats::median, 
                 "std.dev" = stats::sd))

bpFull <- ggplotSampleMDSWrapBiplots(
    mdsObj = mdsObj,
    extVariableList = extVarList,
    pData = flowCore::pData(fsAll),
    projectionAxes = c(1,2),
    pDataForColour = "group",
    pDataForShape = "type",
    seed = 0)

}
\seealso{
\link{ggplotSampleMDS}, \link{ggplotSampleMDSShepard}, \link{computeMetricMDS}
}
